Move rawTo* conversion methods into reusable trait
