+++ rest/shared/src/main/scala/io/udash/rest/DefaultServerREST.scala
@@ -8,22 +8,12 @@ import scala.concurrent.ExecutionContext
 /** Default REST usage mechanism using [[io.udash.rest.DefaultRESTFramework]]. */
 class DefaultServerREST[ServerRPCType : DefaultRESTFramework.AsRealRPC : DefaultRESTFramework.RPCMetadata : DefaultRESTFramework.ValidREST]
                        (override protected val connector: RESTConnector)(implicit ec: ExecutionContext)
-  extends UsesREST[ServerRPCType] {
+  extends UsesREST[ServerRPCType] with RESTConverters {
 
   override val framework = DefaultRESTFramework
 
   override val remoteRpcAsReal: DefaultRESTFramework.AsRealRPC[ServerRPCType] = implicitly[DefaultRESTFramework.AsRealRPC[ServerRPCType]]
   override val rpcMetadata: framework.RPCMetadata[ServerRPCType] = implicitly[framework.RPCMetadata[ServerRPCType]]
-
-  def rawToHeaderArgument(raw: framework.RawValue): String =
-    stripQuotes(framework.rawToString(raw))
-  def rawToQueryArgument(raw: framework.RawValue): String =
-    stripQuotes(framework.rawToString(raw))
-  def rawToURLPart(raw: framework.RawValue): String =
-    URLEncoder.encode(stripQuotes(framework.rawToString(raw)))
-
-  private def stripQuotes(s: String): String =
-    s.stripPrefix("\"").stripSuffix("\"")
 }
 
 object DefaultServerREST {
